Rename common.device -> common.accelerator, return 'gpu' not 'cuda' if torch.cuda.is_available
