+++ src/vak/predict/frame_classification.py
@@ -16,7 +16,7 @@ from tqdm import tqdm
 
 from .. import datasets, models, transforms
 from ..common import constants, files, validators
-from ..common.device import get_default as get_default_device
+from ..common.accelerator import get_default as get_default_device
 from ..datasets.frame_classification import FramesDataset
 
 logger = logging.getLogger(__name__)
