+++ src/vak/predict/predict_.py
@@ -8,7 +8,7 @@ import pathlib
 
 from .. import models
 from ..common import validators
-from ..common.device import get_default as get_default_device
+from ..common.accelerator import get_default as get_default_device
 from .frame_classification import predict_with_frame_classification_model
 
 logger = logging.getLogger(__name__)
