+++ src/vak/train/frame_classification.py
@@ -14,7 +14,7 @@ import torch.utils.data
 
 from .. import datasets, models, transforms
 from ..common import validators
-from ..common.device import get_default as get_default_device
+from ..common.accelerator import get_default as get_default_device
 from ..common.trainer import get_default_trainer
 from ..datasets.frame_classification import FramesDataset, WindowDataset
